refactor: moving utils into seperate repo
