+++ libs/behavioral/package.json
@@ -29,6 +29,6 @@
   },
   "types": "./dist/index.d.ts",
   "dependencies": {
-    "@plaited/utils": "5.0.5"
+    "@plaited/utils": "5.1.0"
   }
 }
