+++ libs/component/package.json
@@ -39,7 +39,7 @@
   },
   "devDependencies": {
     "@plaited/rite": "5.0.4",
-    "@plaited/utils": "5.0.5",
+    "@plaited/utils": "5.1.0",
     "@web/dev-server": "0.4.0",
     "@web/dev-server-esbuild": "1.0.0",
     "@web/test-runner": "0.18.0",
