+++ libs/component/src/index.ts
@@ -3,3 +3,4 @@ export * from './types.js'
 export * from './use-indexed-db.js'
 export * from './use-main.js'
 export * from './use-messenger.js'
+export * from './use-store.js'
