+++ libs/jsx/package.json
@@ -34,6 +34,6 @@
   },
   "types": "./index.d.ts",
   "dependencies": {
-    "@plaited/utils": "5.0.5"
+    "@plaited/utils": "5.1.0"
   }
 }
