--- libs/plaited/src/utils.ts
@@ -1,1 +1,0 @@
-export * from '@plaited/utils'
