--- libs/utils/CHANGELOG.md
@@ -1,110 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## 5.0.5 (2023-11-10)
-
-
-### Bug Fixes
-
-* **component:** delegate listener one weak map per component ([#109](https://github.com/plaited/plaited/issues/109)) ([79a021f](https://github.com/plaited/plaited/commit/79a021fe7669d4ee5318c0ea8c0d104e805cb07e))
-
-
-
-
-
-## 5.0.4 (2023-11-09)
-
-
-### Bug Fixes
-
-* **component:** always delegate events on interactive component host ([#107](https://github.com/plaited/plaited/issues/107)) ([a71baa6](https://github.com/plaited/plaited/commit/a71baa6c5628cd182fe836321c30afec7879e738))
-
-
-
-
-
-## 5.0.3 (2023-11-06)
-
-
-### Bug Fixes
-
-* **jsx:** kebabCase camelcase attributes ([#106](https://github.com/plaited/plaited/issues/106)) ([cb799a6](https://github.com/plaited/plaited/commit/cb799a61a7ca0068628ae56eda345671651f7cfb))
-
-
-
-
-
-## 5.0.2 (2023-11-06)
-
-**Note:** Version bump only for package @plaited/utils
-
-
-
-
-
-## 5.0.1 (2023-11-01)
-
-**Note:** Version bump only for package @plaited/utils
-
-
-
-
-
-# 5.0.0 (2023-10-31)
-
-**Note:** Version bump only for package @plaited/utils
-
-
-
-
-
-# 5.0.0-next-1 (2023-10-31)
-
-**Note:** Version bump only for package @plaited/utils
-
-
-
-
-
-# 4.1.0 (2023-10-28)
-
-**Note:** Version bump only for package @plaited/utils
-
-## 4.0.1 (2023-10-28)
-
-**Note:** Version bump only for package @plaited/utils
-
-# 4.0.0 (2023-10-25)
-
-**Note:** Version bump only for package @plaited/utils
-
-## 3.3.1 (2023-10-25)
-
-**Note:** Version bump only for package @plaited/utils
-
-# 3.3.0 (2023-10-14)
-
-**Note:** Version bump only for package @plaited/utils
-
-# 3.2.0 (2023-07-03)
-
-**Note:** Version bump only for package @plaited/utils
-
-# 3.1.0 (2023-06-03)
-
-**Note:** Version bump only for package @plaited/utils
-
-# Change Log
-
-All notable changes to this project will be documented in this file. See
-[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## 3.0.1 (2023-05-10)
-
-**Note:** Version bump only for package @plaited/utils
-
-# 3.0.0 (2023-05-09)
-
-**Note:** Version bump only for package @plaited/utils
