--- libs/utils/README.md
@@ -1,5 +1,0 @@
-# @plaited/utils
-
-DISCLAIMER: THIS IS AN INTERNAL DEPENDENCY FOR PLAITED LIBS AND APPS. IT IS OPEN SOURCED AS A REFERENCE IMPLEMENTATION. TAKE AS A DEPENDENCY AT YOUR OWN RISK.
-
-Common runtime agnostic utility functions and types used by plaited libraries.
