--- libs/utils/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "@plaited/utils",
-  "version": "5.0.5",
-  "license": "ISC",
-  "engines": {
-    "node": ">= v18.15.0"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/plaited/plaited.git"
-  },
-  "bugs": {
-    "url": "https://github.com/plaited/plaited/issues"
-  },
-  "homepage": "https://github.com/plaited/plaited/tree/main#readme",
-  "type": "module",
-  "main": "./dist/index.js",
-  "exports": {
-    ".": "./dist/index.js"
-  },
-  "files": [
-    "./dist/**"
-  ],
-  "scripts": {
-    "build": "tsc --project tsconfig.json",
-    "build:watch": "bun run build --watch",
-    "test": "bun test src/**/*.spec.ts",
-    "test:watch": "bun run test --watch"
-  },
-  "types": "./dist/index.d.ts"
-}
