--- libs/utils/src/noop.ts
@@ -1,3 +1,0 @@
-/** no-op function good for when you need defaults and stubs */
-// eslint-disable-next-line @typescript-eslint/no-empty-function, @typescript-eslint/no-unused-vars
-export const noop = <T = never>(..._: T[]) => {}
