--- libs/utils/src/reduce-whitespace.ts
@@ -1,2 +1,0 @@
-/** function reduces whitespace in a string down to single spaces */
-export const reduceWhitespace = (str: string) => str.replace(/(\s\s+|\n)/g, ' ')
