--- libs/utils/src/tests/call-all.spec.ts
@@ -1,12 +1,0 @@
-import sinon from 'sinon'
-import { test, expect } from 'bun:test'
-import { callAll } from '../index.js'
-
-test('callAll()', () => {
-  const expected = 'string'
-  const firstSpy = sinon.spy()
-  const secondSpy = sinon.spy()
-  callAll(firstSpy, secondSpy)(expected)
-  expect(firstSpy.calledWith(expected)).toBe(true)
-  expect(secondSpy.calledWith(expected)).toBe(true)
-})
