--- libs/utils/src/tests/key-mirror.spec.ts
@@ -1,10 +1,0 @@
-import { test, expect } from 'bun:test'
-import { keyMirror } from '../index.js'
-
-test('keyMirror(): return an object of mirrored keys as values', () => {
-  expect(keyMirror('a', 'b', 'c')).toEqual({
-    a: 'a',
-    b: 'b',
-    c: 'c',
-  })
-})
