--- libs/utils/src/tests/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2022",
-    "module": "NodeNext",
-    "declaration": true,
-    "moduleResolution": "nodenext",
-    "allowSyntheticDefaultImports": true,
-    "types": ["bun-types"]
-  }
-}
