--- libs/utils/src/value-of.type.ts
@@ -1,1 +1,0 @@
-export type ValueOf<T> = T[keyof T]
