--- libs/utils/src/wait.ts
@@ -1,2 +1,0 @@
-/** an async function that will wait the given time passed to it in ms */
-export const wait = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms))
