--- libs/utils/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2022",
-    "module": "NodeNext",
-    "declaration": true,
-    "moduleResolution": "NodeNext",
-    "outDir": "dist",
-    "rootDir": "src"
-  },
-  "exclude": ["**/dist", "**/tests/"]
-}
